La till fönstret för lägga till måltiderna man har ätit under dagen, tänkt att det ska lägga upp varje matprodukt man har haft i sin måltid sen mängden, därefter visas det i "overview" fönstret totala mängden man har ätit för dagen, dock sparas måltiderna inte ner, samt så saknas återställningsfunktionen efter dagen är över. Man kan även inte ändra måltiden eller ta bort den ännu.
Lagt till även databasen från livsmedelsverket, så datan är från en trovärdig källa.

+++ Models/Nutrient.cs
@@ -13,6 +13,6 @@ namespace ProjectHercules.Models
         public int NutrientId { get; set; }
         public string NutrientName { get; set; }
 
-        public int NutrientValue { get; set; }
+        public float NutrientValue { get; set; }
     }
 }
